Made sash compile, but not link. Need functions, lots of them

+++ nursery/build-attempt3/userapps/sash/cmd_dd.c
@@ -183,7 +183,7 @@ do_dd(int argc, const char ** argv)
 
 	buf = localBuf;
 
-	if (blockSize > sizeof(localBuf))
+	if ((unsigned)blockSize > sizeof(localBuf))
 	{
 		buf = malloc(blockSize);
